# back-end/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CalmWave API",
    "description": "API para plataforma de processamento de áudio com IA. Documentação completa de todas as rotas disponíveis.",
    "version": "1.0.0",
    "contact": {
      "name": "CalmWave Team",
      "email": "support@calmwave.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "profile_photo_url": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "last_access": { "type": "string", "format": "date-time" },
          "active": { "type": "boolean" },
          "account_type": { "type": "string", "enum": ["free", "premium", "admin"] },
          "settings": {
            "type": "object",
            "properties": {
              "dark_mode": { "type": "boolean" },
              "notifications_enabled": { "type": "boolean" },
              "auto_process_audio": { "type": "boolean" },
              "audio_quality": { "type": "string", "enum": ["low", "normal", "high"] }
            }
          }
        }
      },
      "Audio": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "user_id": { "type": "integer" },
          "filename": { "type": "string" },
          "duration_seconds": { "type": "integer" },
          "size_bytes": { "type": "integer" },
          "recorded_at": { "type": "string", "format": "date-time" },
          "processed": { "type": "boolean" },
          "processing_time_ms": { "type": "integer" },
          "transcribed": { "type": "boolean" },
          "transcription_text": { "type": "string" },
          "favorite": { "type": "boolean" },
          "playlist_id": { "type": "integer", "nullable": true },
          "device_origin": { "type": "string" }
        }
      },
      "Playlist": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "user_id": { "type": "integer" },
          "name": { "type": "string" },
          "color": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "total_audios": { "type": "integer" },
          "order": { "type": "integer" }
        }
      },
...
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" },
                  "account_type": { "type": "string", "enum": ["free", "premium"], "default": "free" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "400": { "description": "Dados inválidos" },
          "409": { "description": "Email já cadastrado" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Fazer login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": { "description": "Credenciais inválidas" },
          "403": { "description": "Conta desativada" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Obter dados do usuário autenticado",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Dados do usuário",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": { "description": "Não autorizado" }
        }
      }
    },
...
    "/audios/play/{audio_id}": {
      "get": {
        "tags": ["Audios"],
        "summary": "Reproduzir áudio",
        "parameters": [
          {
            "name": "audio_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          },
          {
            "name": "type",
            "in": "query",
            "schema": { "type": "string", "enum": ["original", "processed"], "default": "processed" }
          }
        ],
        "responses": {
          "200": {
            "description": "Fluxo de áudio",
            "content": {
              "audio/wav": {}
            }
          }
        }
      }
    },
    "/playlists": {
      "get": {
        "tags": ["Playlists"],
        "summary": "Listar playlists do usuário",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "per_page",
            "in": "query",
            "schema": { "type": "integer", "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de playlists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "playlists": { "type": "array", "items": { "$ref": "#/components/schemas/Playlist" } },
                    "total": { "type": "integer" },
                    "pages": { "type": "integer" },
                    "page": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Playlists"],
        "summary": "Criar nova playlist",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "color": { "type": "string", "default": "#6FAF9E" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Playlist criada com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Playlist" }
              }
            }
          }
        }
      }
    },
    "/playlists/{playlist_id}": {
      "get": {
        "tags": ["Playlists"],
        "summary": "Obter playlist específica",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da playlist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "playlist": {
                      "allOf": [
                        { "$ref": "#/components/schemas/Playlist" },
                        {
                          "type": "object",
                          "properties": {
                            "audios": { "type": "array", "items": { "$ref": "#/components/schemas/Audio" } }
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Playlists"],
        "summary": "Atualizar playlist",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "color": { "type": "string" },
                  "order": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Playlist atualizada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Playlist" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Playlists"],
        "summary": "Deletar playlist",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Playlist deletada com sucesso" }
        }
      }
    },
...
  }
}